A queue name must start with a letter or number, and can only contain letters, numbers, and the dash (-) character.
The first and last letters in the queue name must be alphanumeric. The dash (-) character cannot be the first or last character. Consecutive dash characters are not permitted in the queue name.

All letters in a queue name must be lowercase.

A queue name must be from 3 through 63 characters long.